requirements.txt
# instal list
pandas
numpy
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2
python-dotenv